refactor(tests): split tool tests per module and share a parallel pool
tests/tools/ now mirrors the tools modules (annotation, snapshot,
inspect, presenters, workspace, parallel) instead of one grab-bag file;
the snap runner's worker pool moves to tools/parallel.ts so other TS
drivers can fan out through the same primitive.
